Split code into easier to manage files and add reusable functions
