+++ image/remove_pod_affinity.go
@@ -1,0 +1,11 @@
+package main
+
+// removePodAntiAffinity performs the mutation(s) needed to remove podAntiAffinity
+func removePodAntiAffinity(basePath string) (patch []patchOperation) {
+	patch = append(patch, patchOperation{
+		Op:   "remove",
+		Path: basePath,
+	})
+
+	return patch
+}
